--- src/compas_model/algorithms/__init__.py
@@ -1,13 +1,0 @@
-from .nnbrs import find_nearest_neighbours
-
-from .bvh import AABBNode
-from .bvh import OBBNode
-from .bvh import BVH
-
-
-__all__ = [
-    "AABBNode",
-    "OBBNode",
-    "BVH",
-    "find_nearest_neighbours",
-]
