+++ src/compas_model/datastructures/__init__.py
@@ -1,0 +1,14 @@
+from .kdtree import KDTree
+
+from .bvh import (
+    AABBNode,
+    OBBNode,
+    BVH,
+)
+
+__all__ = [
+    "AABBNode",
+    "BVH",
+    "KDTree",
+    "OBBNode",
+]
