+++ src/compas_model/datastructures/bvh.py
